Separate CounterPattern into counter-pattern.cpp
