+++ Makefile.am
@@ -23,7 +23,8 @@ fast_repo_SOURCES = src/main.cpp \
     src/handles/delete-handle.cpp \
     src/handles/read-handle.cpp \
     src/handles/watch-handle.cpp \
-    src/handles/write-handle.cpp
+    src/handles/write-handle.cpp \
+    src/handles/patterns/counter-pattern.cpp
 
 fast_repo_LDADD = ${LIBS} -lcnl-cpp \
     ${BOOST_SYSTEM_LIB} ${BOOST_ASIO_LIB} ${BOOST_THREAD_LIB} ${BOOST_TIMER_LIB}
